Moved VolumeUiState out of the VolumeViewModel
