+++ audio-ui/src/main/java/com/google/android/horologist/audio/ui/VolumePositionIndicator.kt
@@ -37,10 +37,9 @@ import kotlinx.coroutines.delay
  * for 2 seconds after any volume change, including when a new
  * output device is selected.
  */
-@OptIn(ExperimentalHorologistAudioUiApi::class)
 @Composable
 public fun VolumePositionIndicator(
-    volumeUiState: () -> VolumeViewModel.VolumeUiState,
+    volumeUiState: () -> VolumeUiState,
     modifier: Modifier = Modifier,
     autoHide: Boolean = true
 ) {
