+++ media-ui/src/test/java/com/google/android/horologist/media/ui/MediaPlayerTestCase.kt
@@ -39,7 +39,7 @@ import androidx.wear.compose.material.TimeSource
 import androidx.wear.compose.material.TimeText
 import com.google.android.horologist.audio.VolumeState
 import com.google.android.horologist.audio.ui.VolumePositionIndicator
-import com.google.android.horologist.audio.ui.VolumeViewModel
+import com.google.android.horologist.audio.ui.VolumeUiState
 import com.google.android.horologist.audio.ui.components.SettingsButtons
 import com.google.android.horologist.audio.ui.components.SettingsButtonsDefaults
 import com.google.android.horologist.compose.pager.PagerScreen
@@ -107,7 +107,7 @@ fun MediaPlayerTestCase(
                     positionIndicator = {
                         VolumePositionIndicator(
                             volumeUiState = {
-                                VolumeViewModel.VolumeUiState(volumeState = VolumeState(6, 10))
+                                VolumeUiState(volumeState = VolumeState(6, 10))
                             }
                         )
                     }
